(feat): se añade logica a todas las cartas

+++ src/main/java/com/erp/erp_backend/quality/application/request/save/SaveCardPRequest.java
@@ -25,5 +25,5 @@ public class SaveCardPRequest {
     @JsonProperty("proportion")
     @NotNull(message = "proportion is required")
     @Positive(message = "proportion must be greater than 0")
-    private Long proportion;
+    private Double proportion;
 }
