+++ src/main/java/com/erp/erp_backend/quality/infrastructure/entity/CardP.java
@@ -18,7 +18,7 @@ public class CardP {
     private Long lotSize;
     @Column(name = "defective_items")
     private Long defectiveItems;
-    private Long proportion;
+    private Double proportion;
     @Column(name = "user_id")
     private Long userId;
 }
